Silmeden önce "Emin misin?" sorusu eklendi

+++ todo-app/src/components/TodoList.js
@@ -12,7 +12,9 @@ const TodoList = () => {
     }
 
     function deleteClick(itemId){
-        dispatch(deleteTodo(itemId));
+        if(window.confirm("Emin misin?")){
+            dispatch(deleteTodo(itemId));
+        }
     }
 
     return (
